+++ doc/changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Add this changelog to the doc directory
 - Nunique class
+- Add Mode class (supported by a BiCounter, a bi-directional counter)
 
 ### Changed